deploy_temp/bi-engine/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pandas==2.2.3
python-dotenv==1.0.1
python-multipart==0.0.9
openpyxl>=3.1.5
xlrd>=2.0.1
openai>=1.55.0
httpx>=0.27.0
